example-1/requirements.txt
openai
langchain
tiktoken
langchain-openai
faiss-cpu
pypdf
joblib
pdfplumber
python-docx
scikit-learn
transformers
sentence-transformers